refactor: Shop page 컴포넌트 분리
